tasks

+++ 0x1E-search_algorithms/0-linear.c
@@ -1,18 +1,17 @@
 #include "search_algos.h"
 
 /**
-* linear_search - searches for a value in an array of integers using
-* the Linear search algorithm
-* @array: pointer to the first element of the array to search in
-* @size: number of elements in the array
-* @value: value to search for
-*
-* Return: the first index where value is located, or -1 if value is not present
-* in the array or if array is NULL
-*/
+ * linear_search - searches for a value in an array of
+ * integers using the Linear search algorithm
+ *
+ * @array: input array
+ * @size: size of the array
+ * @value: value to search in
+ * Return: Always EXIT_SUCCESS
+ */
 int linear_search(int *array, size_t size, int value)
 {
-size_t i;
+int i;
 if (array == NULL)
 return (-1);
 for (i = 0; i < size; i++)
